+++ src/pages/api/images/create.ts
@@ -1,10 +1,8 @@
-import { imagesControllerPost } from '@/controllers/imagesController'
+import { imagesControllerPost } from '@/pages/controllers/imagesController'
 import type { NextApiRequest, NextApiResponse } from 'next'
-
-interface Data {}
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<Data>) {
   if (req.method === 'POST') {
-    imagesControllerPost(req, res)
+    await imagesControllerPost(req, res)
   }
 }
